% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVisual.R
\name{basicVisualCFLB}
\alias{basicVisualCFLB}
\title{importFrom stats na.omit}
\usage{
basicVisualCFLB(v, vnam, doEval = TRUE)
}
\arguments{
\item{v}{The variable (vector) to be plotted.}

\item{vnam}{The name of the variable which will appear as the title of the plot.}

\item{doEval}{If TRUE, the plot itself is returned. Otherwise, the function returns
a character string containing standalone R code for producing the plot.}
}
\description{
importFrom stats na.omit
}
