% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Perform a check of potential errors in a data frame}
\usage{
check(v, ...)
}
\arguments{
\item{v}{the vector to be checked}

\item{\dots}{other arguments that are passed on the to checking functions}
}
\value{
An list object of class "checked" summarizing the result from the check of the check. As a minimum the returned class should contain the following elements
\itemize{
  \item{"name"}{The name of the check}
  \item{"description"}{Slightly more information about the check}
  \item{"problem"}{An integer giving an error code. 0 means no potential errors were identified}
  \item{"message"}{A string giving summary information in R markdown format about the results}
}
}
\description{
Runs a set of validation checks to check a vector for potential errors.  performs checking steps according to user input and/or data type of the inputted variable.
}
\examples{

x <- 1:5
check(x)

# Annoyingly coded missing as 99
y <- c(rnorm(100), rep(99, 10))

}
\seealso{
\code{\link{clean}}
}
\author{
Anne H. Petersen \email{ahpe@sund.ku.dk} and Claus Thorn Ekstrom \email{ekstrom@sund.ku.dk}
}
\keyword{misc}

