% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunction.R
\name{checkFunction}
\alias{checkFunction}
\title{document me!}
\usage{
checkFunction(f, description, classes = NULL)
}
\arguments{
\item{f}{XXX function}

\item{description}{XXX}

\item{classes}{XXX}
}
\description{
document me!
}

