% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkResult.R
\name{checkResult}
\alias{checkResult}
\title{Create object of class checkResult}
\usage{
checkResult(ls)
}
\arguments{
\item{ls}{A list with entries \code{$problem} (logical indicating whether
a problem was found), \code{$message} (a character string containing a 
message describing the problem) and \code{$problemValues} (the values
in the checked variables that were marked as problematic). Note that
\code{$message} and \code{$problemValues} can be left empty (i.e. 
\code{""} and \code{NULL}, respectively), if they are not relevant.}
}
\value{
A S3 object of class \code{checkResult}, identical to the inputted
list, \code{ls}, except for its class attribute.
}
\description{
Convert a list resulting from the checks performed in a 
\code{\link{checkFunction}} into a \code{checkResult} object, thereby
supplying it with a \code{print()} method.
}
\seealso{
\code{\link{checkFunction}}
}
