% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{classes}
\alias{classes}
\alias{classes<-}
\title{Extract the contents of the attribute \code{classes}}
\usage{
classes(x)

classes(x) <- value
}
\arguments{
\item{x}{The object for which the \code{classes}
attribute should be extracted.}

\item{value}{New value}
}
\value{
The classes for which \code{x} is intended to be called,
given as a vector of characters.
}
\description{
If the object, \code{x}, is itself of
class \code{\link{checkFunction}}, \code{\link{summaryFunction}}
or \code{\link{visualFunction}}, the contents of \code{x}'s
attribute \code{classes} is returned. Otherwise, \code{NULL} is
returned.
}
\examples{
#Extract the classes of the checkFunction identifyMissing
classes(identifyMissing)

#Extract the classes of the summaryFunction minMax
classes(minMax)

#Extract the classes of the visualFunction basicVisual
classes(basicVisual)

}
