% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMissing.R
\name{countMissing}
\alias{countMissing}
\title{Summary function for missing values}
\usage{
countMissing(v)
}
\arguments{
\item{v}{A variable (vector).}
}
\value{
A list with $feature: "No. missing obs." and
$result: [the number and percentage missing observations].
}
\description{
A summary type function to be called from \code{\link{summarize}}, which counts the
number of missing (NA) values in a variable.
}
\examples{
countMissing(c(1:100, rep(NA, 10)))

}
\seealso{
\code{\link{summarize}}
}

