% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{defaultCharacterChecks}
\alias{defaultCharacterChecks}
\title{Default checks for character variables}
\usage{
defaultCharacterChecks()
}
\value{
A vector of function names.
}
\description{
Default options for which checks to perform on
character type variables in \code{\link{check}} and \code{\link{clean}}.
}
