% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultCharacterDescriptions}
\alias{defaultCharacterDescriptions}
\title{Default description functions for character variables}
\usage{
defaultCharacterDescriptions()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default description functions for character variables.
}
\examples{
defaultCharacterDescriptions()

}
\seealso{
\code{\link{defaultCharacterSummaries}}, \code{\link{centralValue}}
}

