% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultCharacterSummaries}
\alias{defaultCharacterSummaries}
\title{Default summary functions for character variables}
\usage{
defaultCharacterSummaries()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default summary functions for character variables.
}
\examples{
defaultCharacterSummaries()

}
\seealso{
\code{\link{variableType}}, \code{\link{countMissing}}, \code{\link{uniqueValues}},
\code{\link{centralValue}}
}
