% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{defaultDateChecks}
\alias{defaultDateChecks}
\title{Default checks for Date variables}
\usage{
defaultDateChecks()
}
\value{
A vector of function names.
}
\description{
Default options for which checks to perform on
Date type variables in \code{\link{check}} and \code{\link{clean}}.
}
