% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultDateDescriptions}
\alias{defaultDateDescriptions}
\title{Default description functions for Date variables}
\usage{
defaultDateDescriptions()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default description functions for Date variables.
}
\examples{
defaultDateDescriptions()

}
\seealso{
\code{\link{defaultDateSummaries}},  \code{\link{minMax}}
}

