% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultDateSummaries}
\alias{defaultDateSummaries}
\title{Default summary functions for Date variables}
\usage{
defaultDateSummaries()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default summary functions for Date variables.
}
\examples{
defaultDateSummaries()

}
\seealso{
\code{\link{variableType}}, \code{\link{countMissing}}, \code{\link{uniqueValues}},
\code{\link{centralValue}}, \code{\link{minMax}}, \code{\link{quartiles}}
}
