% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultFactorDescriptions}
\alias{defaultFactorDescriptions}
\title{Default description functions for factor variables}
\usage{
defaultFactorDescriptions()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default description functions for factor variables.
}
\examples{
defaultFactorDescriptions()

}
\seealso{
\code{\link{defaultFactorSummaries}}, \code{\link{centralValue}}
}

