% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultFactorSummaries}
\alias{defaultFactorSummaries}
\title{Default summary functions for factor variables}
\usage{
defaultFactorSummaries()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default summary functions for factor variables.
}
\examples{
defaultFactorSummaries()

}
\seealso{
code{\link{variableType}}, \code{\link{countMissing}}, \code{\link{uniqueValues}},
\code{\link{centralValue}}
}
