% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{defaultIntegerChecks}
\alias{defaultIntegerChecks}
\title{Default checks for integer variables}
\usage{
defaultIntegerChecks()
}
\value{
A vector of function names.
}
\description{
Default options for which checks to perform on
integer type variables in \code{\link{check}} and \code{\link{clean}}.
}
