% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultIntegerDescriptions}
\alias{defaultIntegerDescriptions}
\title{Default description functions for integer variables}
\usage{
defaultIntegerDescriptions()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default description functions for integer variables.
}
\examples{
defaultIntegerDescriptions()

}
\seealso{
\code{\link{defaultIntgerSummaries}}, \code{\link{centralValue}},
\code{\link{quartiles}}, \code{\link{minMax}}
}

