% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultIntegerSummaries}
\alias{defaultIntegerSummaries}
\title{Default summary functions for integer variables}
\usage{
defaultIntegerSummaries()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default summary functions for integer variables.
}
\examples{
defaultIntegerSummaries()

}
\seealso{
\code{\link{variableType}},
\code{\link{countMissing}}, \code{\link{uniqueValues}},
\code{\link{centralValue}}, \code{\link{quartiles}}, \code{\link{minMax}}
}
