% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{defaultLabelledChecks}
\alias{defaultLabelledChecks}
\title{Default checks for labelled variables}
\usage{
defaultLabelledChecks()
}
\value{
A vector of function names.
}
\description{
Default options for which checks to perform on
labelled type variables in \code{\link{check}} and \code{\link{clean}}.
}
