% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultLabelledDescriptions}
\alias{defaultLabelledDescriptions}
\title{Default description functions for labelled variables}
\usage{
defaultLabelledDescriptions()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default description functions for labelled variables.
}
\examples{
defaultLabelledDescriptions()

}
\seealso{
\code{\link{defaultLabelledSummaries}}, \code{\link{centralValue}}
}

