% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultLabelledSummaries}
\alias{defaultLabelledSummaries}
\title{Default summary functions for labelled variables}
\usage{
defaultLabelledSummaries()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default summary functions for labelled variables.
}
\examples{
defaultLabelledSummaries()

}
\seealso{
\code{\link{variableType}},
\code{\link{countMissing}}, \code{\link{uniqueValues}}, \code{\link{centralValue}}
}
