% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{defaultLogicalChecks}
\alias{defaultLogicalChecks}
\title{Default checks for logical variables}
\usage{
defaultLogicalChecks()
}
\value{
A vector of function names.
}
\description{
Default options for which checks to perform on
logical type variables in \code{\link{check}} and \code{\link{clean}}.
}
