% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultLogicalDescriptions}
\alias{defaultLogicalDescriptions}
\title{Default description functions for logical variables}
\usage{
defaultLogicalDescriptions()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default description functions for logical variables.
}
\examples{
defaultLogicalDescriptions()

}
\seealso{
\code{\link{defaultLogicalSummaries}}, \code{\link{centralValue}}
}

