% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultLogicalSummaries}
\alias{defaultLogicalSummaries}
\title{Default summary functions for logical variables}
\usage{
defaultLogicalSummaries()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default summary functions for logical variables.
}
\examples{
defaultLogicalSummaries()

}
\seealso{
\code{\link{variableType}},
\code{\link{countMissing}}, \code{\link{uniqueValues}}, \code{\link{centralValue}}
}
