% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{defaultNumericChecks}
\alias{defaultNumericChecks}
\title{Default checks for numeric variables}
\usage{
defaultNumericChecks()
}
\value{
A vector of function names.
}
\description{
Default options for which checks to perform on
numeric type variables in \code{\link{check}} and \code{\link{clean}}.
}
