% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultNumericDescriptions}
\alias{defaultNumericDescriptions}
\title{Default description functions for numeric variables}
\usage{
defaultNumericDescriptions()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default description functions for numeric variables.
}
\examples{
defaultNumericDescriptions()

}
\seealso{
\code{\link{defaultNumericSummaries}}, \code{\link{centralValue}},
\code{\link{quartiles}}, \code{\link{minMax}}
}

