% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultNumericSummaries}
\alias{defaultNumericSummaries}
\title{Default summary functions for numeric variables}
\usage{
defaultNumericSummaries()
}
\value{
A list of function names (as character strings).
}
\description{
Returns vector of names of default summary functions for numeric variables.
}
\examples{
defaultNumericSummaries()

}
\seealso{
\code{\link{variableType}},
\code{\link{countMissing}}, \code{\link{uniqueValues}},
\code{\link{centralValue}}, \code{\link{quartiles}}, \code{\link{minMax}}
}
