% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\alias{description<-}
\title{Extract the contents of the attribute \code{description}}
\usage{
description(x)

description(x) <- value
}
\arguments{
\item{x}{The object for which the \code{description}
attribute should be extracted.}

\item{value}{New value}
}
\value{
A description of what \code{x} does, given as
a character string.
}
\description{
If the object, \code{x}, is itself of
class \code{\link{checkFunction}}, \code{\link{summaryFunction}}
or \code{\link{visualFunction}}, the contents of \code{x}'s
attribute \code{description} is returned. Otherwise, \code{NULL} is
returned.
}
\examples{
#Extract the description of the checkFunction identifyMissing
description(identifyMissing)

#Extract the description of the summaryFunction minMax
description(minMax)

#Extract the description of the visualFunction basicVisual
description(basicVisual)

}
