% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyCaseIssues.R
\name{identifyCaseIssues}
\alias{identifyCaseIssues}
\title{A checkFunction for identifying case issues}
\usage{
identifyCaseIssues(v, nMax = Inf)
}
\arguments{
\item{v}{A character or factor variable to check}

\item{nMax}{The maximum number of problematic values to report. Default is \code{Inf}, in which case
all problematic values are included in the outputtet message.}
}
\value{
A list with two elements, $problem: TRUE if any case issues were found, FALSE otherwise, and
$message A message describing which values in \code{v} resulted in case issues. Only unique values
are listed and they are presented in alphabetical order.
}
\description{
A checkFunction to be called from \code{\link{check}} that identifies values in a vector
that appear multiple times with different case settings.
}
\examples{
 identifyCaseIssues(c("val", "b", "1", "1", "vAl", "VAL", "oh", "OH"))

}
\seealso{
\code{\link{check}}, \code{\link{checkFunction}}
}

