% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyMissing.R
\name{identifyMissing}
\alias{identifyMissing}
\title{A checkFunction for identifying miscoded missing values.}
\usage{
identifyMissing(v)
}
\arguments{
\item{v}{A variable to check.}
}
\value{
A list with two elements, $problem: TRUE if any miscoded missing values were found,
FALSE otherwise, and $message A message describing which values in \code{v} were miscoded missing
values. Note that only unique values are listed and that they appear in alphabetical order.
}
\description{
A checkFunction to be called from \code{\link{check}} that identifies values that
appear to be miscoded missing values.
}
\details{
\code{identifyMissing} tries to identify common choices of missing values outside of the
R standards. These include special words (NaN and Inf (no matter the cases)), one or more 9's
(e.g. 999 or "99"), STATA style missing values (commencing with ".") and other character strings
("", " ", "-", "NA" miscoded as character).
}
\examples{
##data(testData)
##testData$miscodedMissingVar
##identifyMissing(testData$miscodedMissingVar)

}
\seealso{
\code{\link{check}}, \code{\link{checkFunction}}
}

