% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyOutliersTBStyle.R
\name{identifyOutliersTBStyle}
\alias{identifyOutliersTBStyle}
\title{A checkFunction for identifying outliers Turkey Boxstole style}
\usage{
identifyOutliersTBStyle(v, nMax = Inf)
}
\arguments{
\item{v}{A character or factor variable to check}

\item{nMax}{The maximum number of problematic values to report. Default is \code{Inf}, in which case
all problematic values are included in the outputtet message.}
}
\value{
A list with two elements, $problem: TRUE if any outliers were found, FALSE otherwise, and
$message A message describing which values in \code{v} were outliers. Note that outlier values
are only printed once (even if they appear multiple times) and that they are ordered.
}
\description{
A checkFunction to be called from \code{\link{check}} that identifies outlier values
in a numeric/integer variable by use of the Turkey Boxplot method (consistent witht the
\code{\link{boxplot}} function).
}
\details{
Outliers are defined in the style of Turkey Boxplots (consistent with the
\code{\link{boxplot}} function), i.e. as values  that are smaller than the 1st quartile minus
the inter quartile range (IQR) or greater than  the third quartile plus the IQR.
}
\examples{
 identifyOutliers(c(1:10, 200, 200, 700))

}
\seealso{
\code{\link{check}}, \code{\link{checkFunction}}
}

