% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyWhitespace.R
\name{identifyWhitespace}
\alias{identifyWhitespace}
\title{A checkFunction for identifying whitespace}
\usage{
identifyWhitespace(v)
}
\arguments{
\item{v}{A character, labelled or factor variable to check.}
}
\value{
A list with two elements, $problem: TRUE if any whitespaces were found, FALSE otherwise, and
$message A message describing which values in \code{v} were prefixed or suffixed with whitespace.
Note that only unique values are printed and that they are sorted alphabetically.
}
\description{
A checkFunction to be called from \code{\link{check}} that identifies prefixed and
suffixed whitespace(s) in character, labelled or factor variables.
}
\examples{
 identifyWhitespace(c("a", " b", "c", "d ", "e  "))

}
\seealso{
\code{\link{check}}, \code{\link{checkFunction}}
}

