% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCPR.R
\name{isCPR}
\alias{isCPR}
\title{Check if a variable consists of Danish CPR numbers}
\usage{
isCPR(v)
}
\arguments{
\item{v}{A variable (vector) to check.}
}
\value{
A list with two elements, $problem: \code{TRUE} if \code{v} consists of CPR numbers,
\code{FALSE} otherwise, and $message A message that either describes the issue (if $problem is \code{TRUE}) 
or is empty (otherwise).
}
\description{
A \code{\link{checkFunction}} that checks if \code{v} consists exclusively of Dansih civil 
registration (CPR) numbers, ignoring missing values.
}
\examples{
CPRs <-  sapply(c("01011988", "02011987", "04052006", "01021990", "01021991",
                  "01021993", "01021994", "01021995", "01021996", "01021997",
                  "01021970", "01021971", "01021972", "01021973", "01021974"), cleanR:::makeCPR)
nonCPRs <- c(1:10)
mixedCPRs <- c(CPRs, nonCPRs)

#identify problem
isCPR(CPRs)

#no problem as there are no CPRs
isCPR(nonCPRs)

#no problem because not ALL values are CPRs
isCPR(mixedCPRs)

}

