% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEmpty.R
\name{isEmpty}
\alias{isEmpty}
\title{Check if a variable only contains a single value}
\usage{
isEmpty(v)
}
\arguments{
\item{v}{A variable (vector) to check.}
}
\value{
A list with two elements, $problem: \code{TRUE} if \code{v} contains only one value,
\code{FALSE} otherwise, and $message A message that either describes the issue (if $problem is \code{TRUE}) 
or is empty (otherwise).
}
\description{
A \code{\link{checkFunction}} that checks if \code{v} only contains a single unique value 
(possibly \code{NA}).NOT TRUE - UPDATE TEXT TO REFLECT WHAT WE DO WITH NAs
}
\examples{
emptyVar <- rep("a", 10)
notEmptyVar <- c("a", "a", "b", "c", "d", "e", "f")

isEmpty(emptyVar)
isEmpty(notEmptyVar)
}

