% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isKey.R
\name{isKey}
\alias{isKey}
\title{Check if a variable is a key in the dataset}
\usage{
isKey(v)
}
\arguments{
\item{v}{A variable (vector) to check.}
}
\value{
A list with two elements, $problem: \code{TRUE} if \code{v} is a key,
\code{FALSE} otherwise, and $message A message that either describes the issue (if $problem is \code{TRUE}) 
or is empty (otherwise).
}
\description{
A \code{\link{checkFunction}} that checks if \code{v} is a key, that is if every 
observation has a unique value in \code{v} and \code{v} is either a factor or character variable.
}
\details{
Note that numeric or integer variables are not considered candidates for keys, as truly
continuous measurements will most likely result in unique values for each observation.
}
\examples{
keyVar <- c("a", "b", "c", "d", "e", "f")
notKeyVar <- c("a", "a", "b", "c", "d", "e", "f")

isKey(keyVar)
isKey(notKeyVar)

}

