% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCodebook.R
\name{makeCodebook}
\alias{makeCodebook}
\title{Produce a data codebook}
\usage{
makeCodebook(data, vol = "", reportTitle = NULL, ...)
}
\arguments{
\item{data}{The dataset to be checked. This dataset should be of class \code{data.frame},
\code{tibble} or \code{matrix}. If it is of classs \code{matrix}, it will be converted to a
\code{data.frame}.}

\item{vol}{Extra text string or numeric that is appended on the end of the output
file name(s). For example, if the dataset is called "myData", no file argument is
 supplied and \code{vol=2}, the output file will be called "codebook_myData2.Rmd"}

\item{reportTitle}{A text string. If supplied, this will be the printed title of the
report. If left unspecified, the title with the name of the supplied dataset.}

\item{...}{Additional parameters passed to \code{makeDataReport}.}
}
\description{
Make a data codebook that summarizes the contents of a dataset.
The result is saved to an R markdown file which can
rendered into an easy-to-read codebook in pdf, html or word formats.
}
