% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minMax.R
\name{minMax}
\alias{minMax}
\title{Description function for minimum and maximum}
\usage{
minMax(v, maxDecimals = 2)
}
\arguments{
\item{v}{A variable (vector) of type numeric or integer.}

\item{maxDecimals}{Number of decimals used when printing numerical values in the data 
summary. If \code{Inf}, no rounding is performed.}
}
\value{
A list with $feature: "Min. and max." and $result: [the minimum and maximum of \code{v}].
}
\description{
A description type function to be called from \code{\link{summarize}}, which returns
the minimum and maximum value of a variable. NA, NaN and Inf values are removed prior to  the
computations.
}
\examples{
minMax(c(1:100))

}
\seealso{
\code{\link{summarize}}
}

