% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Simplified Rmarkdown rendering}
\usage{
render(file, quiet)
}
\arguments{
\item{file}{A character string path to the file that is to be rendered.
This file must be of type Rmarkdown (.Rmd)}

\item{quiet}{A logical. Should messages during rendering be surpressed?}
}
\description{
Render a Rmarkdown (.Rmd) file, \code{file}, to the output
format specified in its preamble. If no output format is specified,
it will be rendered to html.
}
\details{
This function is merely a simplified version (in terms of
possible arguments) of the rendering function from the \code{rmarkdown} package.
Therefore, we refer to this functions for more details:
\code{\link[rmarkdown]{render}}. We have included this simplified version in
\code{dataMaid} in order to help new R users with rendering their output
documents as generated by \code{\link{clean}}.
}
\seealso{
\code{\link[rmarkdown]{render}}.
}
