% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryResult.R
\name{summaryResult}
\alias{summaryResult}
\title{Create object of class summaryResult}
\usage{
summaryResult(ls)
}
\arguments{
\item{ls}{A list with entries \code{$feature} (a character string describing
what summary was obtained), \code{$result} (the result of the summary, either
a value from the variable, a numeric or a character string) and 
\code{$value} (the result in its most raw format, often identical to the 
\code{$result} input).}
}
\value{
A S3 object of class \code{summaryResult}, identical to the inputted
list, \code{ls}, except for its class attribute.
}
\description{
Convert a list resulting from the summaries performed in a
\code{\link{summaryFunction}} into a \code{summaryResult} object, thereby
supplying it with a \code{print()} method.
}
\seealso{
\code{\link{summaryFunction}}
}
