% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMaid-package.R
\docType{data}
\name{toyData}
\alias{toyData}
\title{Small example data to show the features of dataMaid}
\format{A \code{data.frame} with 15 rows and 6 variables.
\describe{
   \item{var1}{A factor variable with two levels (\code{"red"} and \code{"blue"}) and a few
   (correctly coded)  missing observations.}
   \item{var2}{A numeric variable with one obvious outlier value (\code{82}), two miscoded
   missing values (\code{999} and \code{NaN}) and a few correctly coded missing values.}
   \item{var3}{A factor variable where two of the levels (\code{"other"} and \code{"OTHER"}
   are the same word with different case settings. Moreover, the variable includes a Stata-style
   miscoded missing value (\code{"."})}.
   \item{var4}{A numeric variable (random draws from a standard normal distribution).}
   \item{var5}{A factor variable with unique codes for each observation (a character string
   with a number between 1 and 15), i.e. a key variable.}
   \item{var6}{A factor variable that has the same level (\code{"Irrelevant"}) for all
   observations, i.e. a empty variable.}
}}
\source{
Artificial data
}
\usage{
toyData
}
\description{
An artificial dataset, intended for presenting the key features of \code{dataMaid}, which is a
toolset for identifying potential errors in a dataset.
}
\examples{
data(toyData)

}
\keyword{datasets}
