% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueValues.R
\name{uniqueValues}
\alias{uniqueValues}
\title{Summary function for unique values}
\usage{
uniqueValues(v)
}
\arguments{
\item{v}{A variable (vector).}
}
\value{
A list with $feature: "No. unique values" and
$result: [the number of unique values].
}
\description{
A summary type function to be called from \code{\link{summarize}}, which counts the
number of unique (excluding NA) values in a variable.
}
\examples{
uniqueValues(c(1:3, rep(NA, 10), Inf, NaN))

}
\seealso{
\code{\link{summarize}}
}

