% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableType.R
\name{variableType}
\alias{variableType}
\title{Summary function for original class}
\usage{
variableType(v, ...)
}
\arguments{
\item{v}{A variable (vector).}

\item{...}{Not in use.}
}
\value{
A list with $feature: "Variable type" and
$result: [the (original) class of the variable].
}
\description{
A summary type function to be called from \code{\link{summarize}}, which finds the
orignial class of a variable. This is just the class for all objects but those of class
smartNum.
}
\examples{
 #For standard variables:
   varX <- c(rep(c(1,2,3), each=10))
   class(varX)
   variableType(varX)

 #For smartNum variables:
   smartX <- smartNum(varX)
   class(smartX)
   variableType(smartX)

 #Something with multiple classes, like ordered factor...

}
\seealso{
\code{\link{summarize}}
}

