% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{whoami_available}
\alias{whoami_available}
\title{Find out if the whoami package binaries is installed (git + whoami)}
\usage{
whoami_available()
}
\value{
logical that is TRUE if whoami and git can be found
}
\description{
Find out if the whoami package binaries is installed (git + whoami)
}
