% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{whoami_exists}
\alias{whoami_exists}
\title{Find path to the pandoc binary by checking the \code{PATH} and the \code{RSTUDIO_PANDOC} env vars}
\usage{
whoami_exists()
}
\value{
logical that is TRUE is pandoc can be found
}
\description{
Find path to the pandoc binary by checking the \code{PATH} and the \code{RSTUDIO_PANDOC} env vars
}
