% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralValue.R
\name{centralValue}
\alias{centralValue}
\title{Description function for central values}
\usage{
centralValue(v)
}
\arguments{
\item{v}{A variable (vector).}
}
\value{
A list with $feature: [Mode/median] and $result: [the central value of \code{v}].
}
\description{
A description type function to be called from \code{\link{summarize}}, which returns
the central value of a variable. For numeric and integer variables, this is the median. For
character, factor, labelled and logical variables, the central value is the mode (i.e. the
value that occurs the largest number of times).
}
\details{
Note that NA, NaN and Inf values are ignored for numeric and integer variables, while
only NA values are ignored for factor, character and labelled variables. No values are
ignored for logical variables.
}
\examples{
 #central value of a integer variable:
   centralValue(c(rep(1, 25), rep(2, 10), rep(3, 20)))

 #central value of a character variable:
   centralValue(as.character(c(rep(1, 20), rep(2, 10), rep(3, 20))))

}
\seealso{
\code{\link{summarize}}
}

