% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Perform a check of potential errors in a data frame}
\usage{
clean(o, file = NULL, removeExisting = TRUE, standAlone = TRUE,
  ordering = c("asIs", "alphabetical"), quiet = TRUE,
  output = c("markdown", "pdf", "html", "screen"), render = TRUE,
  twoCol = TRUE, silent = FALSE, openResult = TRUE,
  mode = c("summarize", "visualize", "check"), useVar = NULL,
  onlyProblematic = FALSE, nagUser = TRUE, smartNum = TRUE,
  preChecks = c("isSpecial", "isCPR"), replace = FALSE, listChecks = TRUE,
  checkDetails = FALSE, garbageCollection = TRUE, vol = "", ...)
}
\arguments{
\item{o}{the data frame object to be checked.}

\item{file}{The filename of output file. If set to NULL (the default) then the filename will be the name of the data frame prefixed with "cleanR-".}

\item{removeExisting}{XXX}

\item{standAlone}{If TRUE, the document begins with a markdown preamble such that it
can be rendered as is.}

\item{ordering}{Choose the ordering of the variables in the data presentation. The options
are "asIs" (ordering as in the dataset) and "alpha" (alphabetical order).}

\item{output}{Output format. Options are "markdown" (the default), "pdf", "html", and "screen".}

\item{render}{Should the output file be rendered (defaults to TRUE)? This argument has no impact for screen output.}

\item{twoCol}{Should the results be presented in two columns (if output is "html" or "pdf")? Defaults to TRUE.}

\item{silent}{Should clean() run completely silently? Note that this option overrules the settings for
"quiet": A silent session is always quiet.}

\item{openResult}{If TRUE, the file produced by clean() is automatically opened by the end of
the function run.}

\item{mode}{Vector of tasks to perform among the three categories "summarize", "visualize" and "check".
Note that... SOMETHING ABOUT HOW THE FUNCTIONS CALLED IN EACH PART ARE CONTROLLED.}

\item{useVar}{Variables to clean. If NULL (the default) then all variables in the data frame o are included. If a vector of variable names is included then only the variables in o that are also part of useVar are checked.}

\item{onlyProblematic}{A logical. Set to TRUE if only the potentially problematic variables should be listed.}

\item{nagUser}{Remove at some point}

\item{smartNum}{If TRUE, numeric and integer variables with less than maxLevels (defaults to 5) unique
values are treated as factor variables in the checking, visualization and summary functions. A
message is printed in the data summary as well.}

\item{preChecks}{Variable checks that are performed before the summary/visualization/checking step. If
any of these checks find problems, the variable will not be summarized nor visualized nor checked.}

\item{replace}{If FALSE (the default) an error is thrown if one of the files that we are about to write to
already exists. If TRUE no checks are performed.}

\item{listChecks}{If TRUE, the document contains an overview of what checks were performed for
each variable data type.}

\item{checkDetails}{MAYBE ALSO IMPLEMENT THIS?: If TRUE, details about each check function are added
to the document (if available)}

\item{garbageCollection}{A logical. If TRUE (the default) then garbage collection code is added to the R markdown file that is output. This is useful for larger dataset to prevent memory problems.}

\item{vol}{Extra text string that is appended on the end of the output file name(s). For example, if the data
set is called "myData", no file argument is supplied and vol="2", the output file will be called
"cleanR_myData2.Rmd"}

\item{characterChecks}{a list of error-checking functions to apply to character vectors}

\item{integerChecks}{a list of error-checking functions to apply to integer vectors}

\item{\dots}{other arguments that are passed on the to checking, summary and visualization functions}
}
\value{
???
}
\description{
Runs a set of validation checks to check the variables in a data frame for potential errors.
Performs checking steps according to user input and/or data type of the inputted variable.
}
\examples{

clean(data.frame(x=1:15))

\dontrun{
data(testData)
clean(testData)
}

#Add user defined check-function to the checks performed on character variables:
\dontrun{
characterFoo <- function(v) {
 if (substr(substitute(a), 1, 1) == "_") {
   out <- list(problem=TRUE, message="Note that the variable name begins with \\\\_")
 } else out <- list(problem=FALSE, message="")
 out
}
class(characterFoo) <- "checkFunction"
attr(characterFoo, "description") <- "I really hate underscores"
clean(testData, characterChecks=c(defaultCharacterChecks(), "characterFoo"))
}

}
\author{
Anne H. Petersen \email{ahpe@sund.ku.dk} and Claus Thorn Ekstrom \email{ekstrom@sund.ku.dk}
}
\seealso{
\code{\link{clean}}
}
\keyword{misc}

