% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Perform a check of potential errors in a data frame}
\usage{
clean(data, output = c("pdf", "html", "screen"), render = TRUE,
  useVar = NULL, ordering = c("asIs", "alphabetical"),
  onlyProblematic = FALSE, labelled_as = c("factor", "NA", "zap"),
  mode = c("summarize", "visualize", "check"), smartNum = TRUE,
  preChecks = c("isKey", "isEmpty"), file = NULL, replace = FALSE,
  vol = "", standAlone = TRUE, twoCol = TRUE, quiet = TRUE,
  openResult = TRUE, nagUser = TRUE, checkDetails = FALSE,
  characterChecks = defaultCharacterChecks(),
  factorChecks = defaultFactorChecks(),
  labelledChecks = defaultLabelledChecks(),
  numericChecks = defaultNumericChecks(),
  integerChecks = defaultIntegerChecks(),
  logicalChecks = defaultLogicalChecks(), dateChecks = defaultDateChecks(),
  allChecks = NULL, characterSummaries = defaultCharacterSummaries(),
  factorSummaries = defaultFactorSummaries(),
  labelledSummaries = defaultLabelledSummaries(),
  numericSummaries = defaultNumericSummaries(),
  integerSummaries = defaultIntegerSummaries(),
  logicalSummaries = defaultLogicalSummaries(),
  dateSummaries = defaultDateSummaries(), allSummaries = NULL,
  allVisuals = "standardVisual", listChecks = TRUE, brag = FALSE,
  maxProbVals = Inf, maxDecimals = 2, ...)
}
\arguments{
\item{data}{the dataset to be checked [LIST ALL ALLOWED OBJECT CLASSES, INCLUDING data.frame,
matrix, data.table, ... (more?)].}

\item{output}{Output format. Options are "markdown" (the default), "pdf", "html", and "screen".
All but the "screen" option produces an R markdown file which can be rendered.
The "screen" option prints a small summary on the screen.}

\item{render}{Should the output file be rendered (defaults to TRUE)? This argument has no
impact unless the output is "html" or "pdf" in which case the R markdown file is rendered to
produce the corresponding file.}

\item{useVar}{Variables to clean. If NULL (the default) then all variables in the data.frame
o are included. If a vector of variable names is included then only the variables in o that are
also part of useVar are checked.}

\item{ordering}{Choose the ordering of the variables in the data presentation. The options
are "asIs" (ordering as in the dataset) and "alphabetical" (alphabetical order).}

\item{onlyProblematic}{A logical. Set to TRUE if only the potentially problematic variables should be listed.}

\item{labelled_as}{A string explaining the way to handle labelled vectors. Possibilities are either "factor" (the default), "NA" (all missing are converted to NA) or "zap" (remove all labels).}

\item{mode}{Vector of tasks to perform among the three categories "summarize", "visualize" and "check".
Note that... SOMETHING ABOUT HOW THE FUNCTIONS CALLED IN EACH PART ARE CONTROLLED.}

\item{smartNum}{If TRUE, numeric and integer variables with less than maxLevels (defaults to 5) unique
values are treated as factor variables in the checking, visualization and summary functions. A
message is printed in the data summary as well.}

\item{preChecks}{Variable checks that are performed before the summary/visualization/checking step. If
any of these checks find problems, the variable will not be summarized nor visualized nor checked.}

\item{file}{The filename of output file. If set to NULL (the default) then the filename will be
the name of the data frame prefixed with "cleanR_". Note that a valid file is of type rmd, hence all
filenames should have a ".Rmd"-suffix.}

\item{replace}{If FALSE (the default) an error is thrown if one of the files that we are about to write to
already exists. If TRUE no checks are performed.}

\item{vol}{Extra text string that is appended on the end of the output file name(s). For example, if the data
set is called "myData", no file argument is supplied and vol="2", the output file will be called
"cleanR_myData2.Rmd"}

\item{standAlone}{If TRUE, the document begins with a markdown preamble such that it
can be rendered as a stand alone R markdown file.}

\item{twoCol}{Should the results be presented in two columns (if output is "html" or "pdf")? Defaults to TRUE.}

\item{quiet}{If \code{TRUE} (the default), only a few messages is printed to the screen as \code{clean} runs.
If \code{FALSE} no messages are suppressed. The third option, \code{silent}, renders the function
completely silent such that only fatal errors are printed.}

\item{openResult}{If TRUE, the file produced by clean() is automatically opened by the end of
the function run.}

\item{nagUser}{Remove at some point}

\item{checkDetails}{MAYBE ALSO IMPLEMENT THIS?: If TRUE, details about each check function are added
to the document (if available)}

\item{characterChecks}{a list of error-checking functions to apply to character vectors}

\item{factorChecks}{a list of error-checking functions to apply to integer vectors}

\item{labelledChecks}{a list of error-checking functions to apply to character vectors}

\item{numericChecks}{a list of error-checking functions to apply to integer vectors}

\item{integerChecks}{a list of error-checking functions to apply to integer vectors}

\item{logicalChecks}{a list of error-checking functions to apply to Date vectors}

\item{dateChecks}{a list of error-checking functions to apply to integer vectors}

\item{allChecks}{Vector of function names that should be used as check-functions for all variable types.
See ???? for more details OR SOMETHING? Note that this option overwrite the options characterChekcs,
factorChecks, etc.}

\item{characterSummaries}{STUFF}

\item{factorSummaries}{STUFF}

\item{labelledSummaries}{STUFF}

\item{numericSummaries}{STUFF}

\item{integerSummaries}{STUFF}

\item{logicalSummaries}{STUFF}

\item{dateSummaries}{STUFF}

\item{allSummaries}{Vector of function names that should be used as summary-functions for all variable types.
See ???? for more details OR SOMETHING? Note that this option overwrites the options charachterSummaries,
factorSummaries, etc.}

\item{allVisuals}{STUFF. Default: "standardVisual".}

\item{listChecks}{Logical. Sets whether the checks that were used for each possible variable type are shown in the output. Defaults to TRUE.}

\item{brag}{A logical that determines if the name of cleanR package is printed in the output. Defaults to TRUE but should probably be removed altogether.}

\item{maxProbVals}{Maximum number of unique values printed from check-functions (integer > 0).
Defaults to \code{Inf}, which means that all values are printed.}

\item{maxDecimals}{Number of decimals used when printing numerical values in the data
summary. If \code{Inf}, no rounding is performed.}

\item{\dots}{other arguments that are passed on the to precheck, checking, summary and visualization functions}
}
\value{
The function does not return anything. It's side effect (the production of the Rmd file summary)
is the reason for running the function.
}
\description{
Runs a set of validation checks to check the variables in a data frame for potential errors.
Performs checking steps according to user input and/or data type of the inputted variable.
The checks are saved to an R markdown file which can rendered into an easy-to-read document.
}
\examples{

 \dontrun{
DF <- data.frame(x = 1:15)
clean(DF)
}

\dontrun{
data(testData)
clean(testData)
}

#Add user defined check-function to the checks performed on character variables:
\dontrun{
characterFoo <- function(v) {
 if (substr(substitute(v), 1, 1) == "_") {
   out <- list(problem=TRUE, message="Note that the variable name begins with \\\\_")
 } else out <- list(problem=FALSE, message="")
 out
}
class(characterFoo) <- "checkFunction"
attr(characterFoo, "description") <- "I really hate underscores"
clean(testData, characterChecks=c(defaultCharacterChecks(), "characterFoo"))
}

}
\seealso{
\code{\link{clean}}
}
\author{
Anne H. Petersen \email{ahpe@sund.ku.dk} and Claus Thorn Ekstrom \email{ekstrom@sund.ku.dk}
}
\keyword{misc}

