% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMissing.R
\name{countMissing}
\alias{countMissing}
\title{Summary function for missing values}
\usage{
countMissing(v, ...)
}
\arguments{
\item{v}{A variable (vector).}

\item{...}{Not in use.}
}
\value{
A \code{\link{summaryResult}} object with the following entries: 
\code{$feature} ("No. missing obs."), \code{$result} (the number and percentage 
missing observations) and \code{$value} (the number of missing observations).
}
\description{
A \code{\link{summaryFunction}}, intended to be called from
\code{\link{summarize}} (and \code{\link{makeDataReport}}), which counts the
number of missing (\code{NA}) values in a variable.
}
\examples{
countMissing(c(1:100, rep(NA, 10)))

}
\seealso{
\code{\link{summarize}}, \code{\link{allSummaryFunctions}},
\code{\link{summaryFunction}}, \code{\link{summaryResult}}
}
