% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{defaultDateChecks}
\alias{defaultDateChecks}
\title{Default checks for Date variables}
\usage{
defaultDateChecks(remove = NULL, add = NULL)
}
\arguments{
\item{remove}{Character vector of function names. Checks to remove from the returned vector}

\item{add}{Character vector of function names. Checks to add to the returned vector}
}
\value{
A vector of function names.
}
\description{
Default options for which checks to perform on
Date type variables in \code{\link{check}} and \code{\link{makeDataReport}},
possibly user-modified by adding extra function names using \code{add} or 
removing default function names with \code{remove}.
}
