% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultIntegerSummaries}
\alias{defaultIntegerSummaries}
\title{Default summary functions for integer variables}
\usage{
defaultIntegerSummaries(remove = NULL, add = NULL)
}
\arguments{
\item{remove}{Character vector of function names. Checks to remove from the returned vector}

\item{add}{Character vector of function names. Checks to add to the returned vector}
}
\value{
A list of function names (as character strings).
}
\description{
Default options for which summaries to apply on
integer type variables in \code{\link{check}} and \code{\link{makeDataReport}},
possibly user-modified by adding extra function names using \code{add} or 
removing default function names with \code{remove}.
}
\examples{
#remove "countMissing":
defaultIntegerSummaries(remove = "countMissing")

}
\seealso{
\code{\link{variableType}},
\code{\link{countMissing}}, \code{\link{uniqueValues}},
\code{\link{centralValue}}, \code{\link{quartiles}}, \code{\link{minMax}}
}
