% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{defaultLogicalSummaries}
\alias{defaultLogicalSummaries}
\title{Default summary functions for logical variables}
\usage{
defaultLogicalSummaries(remove = NULL, add = NULL)
}
\arguments{
\item{remove}{Character vector of function names. Checks to remove from the returned vector}

\item{add}{Character vector of function names. Checks to add to the returned vector}
}
\value{
A list of function names (as character strings).
}
\description{
Default options for which summaries to apply on
logical type variables in \code{\link{check}} and \code{\link{makeDataReport}},
possibly user-modified by adding extra function names using \code{add} or 
removing default function names with \code{remove}.
}
\examples{
#remove "uniqueValues":
defaultLogicalSummaries(remove = "uniqueValues")

}
\seealso{
\code{\link{variableType}},
\code{\link{countMissing}}, \code{\link{uniqueValues}}, \code{\link{centralValue}}
}
