% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyLoners.R
\name{identifyLoners}
\alias{identifyLoners}
\title{A checkFunction for identifying sparsely represented values (loners)}
\usage{
identifyLoners(v)
}
\arguments{
\item{v}{A character or factor variable to check.}
}
\value{
A list with two elements, $problem: TRUE if any loners were found, FALSE otherwise, and
$message A message describing which values in \code{v} were loners. Note that loner values
are listed only once (no matter how many time they occur) in alphabetical order.
}
\description{
A checkFunction to be called from \code{\link{check}} that identifies values that
only occur less than 6 times in factor or character variables (that is, loners).
}
\examples{
identifyLoners( c(rep(c("a", "b", "c"), 10), "d", "d") )

}
\seealso{
\code{\link{check}}, \code{\link{checkFunction}}
}

