% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyMissing.R
\name{identifyMissing}
\alias{identifyMissing}
\title{A checkFunction for identifying miscoded missing values.}
\usage{
identifyMissing(v, nMax = Inf, ...)
}
\arguments{
\item{v}{A variable to check.}

\item{nMax}{The maximum number of problematic values to report. Default is \code{Inf}, in which case
all problematic values are included in the outputtet message.}

\item{...}{Not in use.}
}
\value{
A list with two elements, $problem: TRUE if any miscoded missing values were found,
FALSE otherwise, and $message A message describing which values in \code{v} were miscoded missing
values. Note that only unique values are listed and that they appear in alphabetical order.
}
\description{
A checkFunction to be called from \code{\link{check}} that identifies values that
appear to be miscoded missing values.
}
\details{
\code{identifyMissing} tries to identify common choices of missing values outside of the
R standards. These include special words (NaN and Inf (no matter the cases)), one or more -9/9's
(e.g. 999, "99", -9, "-99"), one ore more -8/8's (e.g. -8, 888, -8888), STATA style missing values
(commencing with ".") and other character strings ("", " ", "-", "NA" miscoded as character). If 
the variable is numeric/integer or a character/factor variable consisting only of numbers and with
more than 11 different values, the numeric miscoded missing values (999, 888, -99, -8 etc.) are 
only recognized as miscoded missing if they are max or min, respectively, and the distance 
between the second largest/smallest value and this max/min value is greater than one.
}
\examples{
##data(testData)
##testData$miscodedMissingVar
##identifyMissing(testData$miscodedMissingVar)

#Identify miscoded numeric missing values
v1 <- c(1:15, 99)
v2 <- c(v1, 98)
v3 <- c(-999, v2, 9999)
identifyMissing(v1) 
identifyMissing(v2) 
identifyMissing(v3)
identifyMissing(factor(v3))

}
\seealso{
\code{\link{check}}, \code{\link{checkFunction}}
}

