% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartiles.R
\name{quartiles}
\alias{quartiles}
\title{Description function for quartiles}
\usage{
quartiles(v, maxDecimals = 2)
}
\arguments{
\item{v}{A variable (vector) of type numeric or integer.}

\item{maxDecimals}{Number of decimals used when printing numerical values in the data 
summary. If \code{Inf}, no rounding is performed.}
}
\value{
A list with $feature: "1st and 3rd quartiles" and $result: [the quartiles of \code{v}].
}
\description{
A description type function to be called from \code{\link{summarize}}, which calculates
the 1st and 3rd quartiles of a variable. NA, NaN and Inf values are removed prior to  the
computations.
}
\examples{
quartiles(c(1:100))

}
\seealso{
\code{\link{summarize}}
}

