% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize a variable/dataset}
\usage{
summarize(v, reportstyleOutput = FALSE, summaries = setSummaries(), ...)
}
\arguments{
\item{v}{The variable (vector) or dataset (data.frame) to be summarized.}

\item{reportstyleOutput}{Logical indicating whether the output should
be formatted for inclusion in the report (escaped matrix) or not. Defaults to not.}

\item{summaries}{A list of summaries to use on each supported variable type. We recommend
using \code{\link{setSummaries}} for creating this list and refer to the documentation
of this function for more details.}

\item{...}{Additional argument passed to data class specific methods.}
}
\value{
The return value depends on the value of \code{reportstyleOutput}. 

If \code{reportstyleOutput = FALSE} (the default): If \code{v} is a varibale, 
a list of \code{summaryResult} objects, one \code{summaryResult} for each summary
function called on \code{v}. If \code{v} is a dataset, then \code{summarize()} returns 
a list of lists of \code{summaryResult} objects instead; one list for each variable
in \code{v}.  

If \code{reportstyleOutput = TRUE}: 
If \code{v} is a single variable: A matrix with two columns, \code{feature} and 
\code{result} and one row for each summary function that was called. Character
strings in this matrix are escaped such that they are ready for Rmarkdown rendering.

If \code{v} is a full dataset: A list of matrices as described above, one for each
variable in the dataset.
}
\description{
Generic shell function that produces a summary of a variable (or for each
variable in an entire dataset), given a number of summary functions and 
depending on its data class.
}
\details{
Summary functions are supplied using their
names (in character strings) in the class-specific argument, e.g.
\code{characterSummaries = c("countMissing", "uniqueValues")} for character variables and
similarly for the remaining 7 data classes (factor, Date, labelled, haven_labelled, numeric, integer, logical).
Note that an overview of all available \code{summaryFunction}s can be obtained by calling
\code{\link{allSummaryFunctions}}. 

The default choices of \code{summaryFunctions} are available in data class specific functions, e.g.
\code{defaultCharacterSummaries()} and \code{defaultNumericSummaries()}. 
A complete overview of all default options can be obtained by calling setSummaries()

A user defined summary function can be supplied using its function name. Note
however that it should take a vector as argument and return a list on the form
\code{list(feature="Feature name", result="The result")}. More details on how to construct 
valid summary functions are found in \code{\link{summaryFunction}}.
}
\examples{
#Default summary for a character vector:
   charV <- c("a", "b", "c", "a", "a", NA, "b", "0")
   summarize(charV)

#Inspect default character summary functions:
   defaultCharacterSummaries()

#Define a new summary function and add it to the summary for character vectors:
   countZeros <- function(v, ...) {
     res <- length(which(v == 0))
     summaryResult(list(feature="No. zeros", result = res, value = res))
   }
   summarize(charV, 
     summaries = setSummaries(character = defaultCharacterSummaries(add = "countZeros")))

 #Does nothing, as intV is not affected by characterSummaries
   intV <- c(0:10)
   summarize(intV, 
     summaries = setSummaries(character = defaultCharacterSummaries(add = "countZeros")))

 #But supplying the argument for integer variables changes the summary:
   summarize(intV, summaries = setSummaries(integer = "countZeros"))
   
 #Summarize a full dataset:
  data(cars)
  summarize(cars)
  
 #Summarize a variable and obtain report-style output (formatted for markdown)
  summarize(charV, reportstyleOutput = TRUE)

}
\references{
Petersen AH, Ekstrøm CT (2019). “dataMaid: Your Assistant for Documenting Supervised Data Quality Screening in R.” _Journal of Statistical Software_, *90*(6), 1-38. doi: 10.18637/jss.v090.i06 (\url{https://doi.org/10.18637/jss.v090.i06}).
}
\seealso{
\code{\link{setSummaries}},
\code{\link{summaryFunction}}, \code{\link{allSummaryFunctions}}, 
\code{\link{summaryResult}},
\code{\link{defaultCharacterSummaries}}, \code{\link{defaultFactorSummaries}},
\code{\link{defaultLabelledSummaries}}, \code{\link{defaultHavenlabelledSummaries}},
\code{\link{defaultNumericSummaries}}, \code{\link{defaultIntegerSummaries}},
\code{\link{defaultLogicalSummaries}}
}
